+++ test/td_dd_web/schema/implementations_test.exs
@@ -1,5 +1,6 @@
 defmodule TdDdWeb.Schema.ImplementationsTest do
   use TdDdWeb.ConnCase
+  import TdDd.TestOperators
 
   @submit_implementation """
   mutation SubmitImplementation($id: ID!) {
@@ -173,7 +174,8 @@ defmodule TdDdWeb.Schema.ImplementationsTest do
                %{"id" => to_string(rule_result_id_1)},
                %{"id" => to_string(rule_result_id_2)},
                %{"id" => to_string(rule_result_id_3)}
-             ] == Map.get(implementation, "results")
+             ]
+             <|> Map.get(implementation, "results")
     end
 
     @tag authentication: [role: "admin"]