+++ lib/td_dd_web/schema/implementations.ex
@@ -60,6 +60,10 @@ defmodule TdDdWeb.Schema.Implementations do
     field :version, :integer
     field :versions, list_of(:implementation), resolve: &Resolvers.Implementations.versions/3
     field :results, list_of(:implementation_result), resolve: &Resolvers.Implementations.results/3
+
+    field :last_quality_event, :quality_event,
+      resolve: &Resolvers.Implementations.last_quality_event/3
+
     field :updated_at, :datetime
     field :deleted_at, :datetime
   end
@@ -117,5 +121,11 @@ defmodule TdDdWeb.Schema.Implementations do
   object :structure_reference do
     field :id, non_null(:id)
     field :parent_index, :integer
+  end
+
+  object :quality_event do
+    field :inserted_at, :datetime
+    field :message, :string
+    field :type, :string
   end
 end