+++ lib/td_dq/implementations/condition_row.ex
@@ -134,8 +134,14 @@ defmodule TdDq.Implementations.ConditionRow do
     end
   end
 
+  # One field
   defp valid_attribute(%{"id" => id}, _params) do
     is_integer(id)
+  end
+
+  # Several fields
+  defp valid_attribute(%{"fields" => fields}, %{"operator" => %{"value_type" => value_type}}) do
+    is_valid_type_value(value_type, fields)
   end
 
   defp valid_attribute(%{"raw" => raw}, %{"operator" => %{"value_type" => value_type}}) do
@@ -168,7 +174,19 @@ defmodule TdDq.Implementations.ConditionRow do
     is_list(value)
   end
 
+  defp is_valid_type_value("field_list", value) do
+    is_list(value) and Enum.all?(value, &is_field(&1))
+  end
+
   defp is_valid_type_value(_other_type, _value) do
+    false
+  end
+
+  defp is_field(%{"id" => id}) do
+    is_integer(id)
+  end
+
+  defp is_field(_) do
     false
   end