+++ CHANGELOG.md
@@ -5,8 +5,8 @@
 ### Added
 
 - [TD-4176]
-  - Added descrition in datas tructure tag
-  - changed desription by comment in data structures tags link
+  - Added descrition in data structure tag
+  - Changed description by comment in data structures tags link
 
 ### Fixed