+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 ## [Unreleased]
 
 ### Changed
@@ -10,6 +11,7 @@
 - [TD-4920] Add results to implementation query api/v2
 - [TD-4921] Add implementation workflow events
 - [TD-4894] Multiple field validations
+- [TD-4993] Add last_quality_event to implementation api/v2 
 
 ### Fixed