+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 ## [Unreleased]
 
 ### Changed
@@ -9,6 +10,8 @@
 
 - [TD-4920] Add results to implementation query api/v2
 - [TD-4921] Add implementation workflow events
+- [TD-4894] Multiple field validations
+- [TD-4993] Add last_quality_event to implementation api/v2 
 
 ### Fixed
 
@@ -18,7 +21,7 @@
 
 ### Fixed
 
-- [TD-4962] Submit results only for published implementations 
+- [TD-4962] Submit results only for published implementations
 
 ## [4.46.0] 2022-06-21